% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vector_corr.R
\name{vector_corr}
\alias{vector_corr}
\title{Correlation coefficient for vectors}
\usage{
vector_corr(W1, W2)
}
\arguments{
\item{W1}{The first set of vectors expressed by their scalar components in a two-column matrix or data frame (u and v, the change along the x and y axes, respectively)}

\item{W2}{The second set of vectors expressed by their scalar components in a two-column matrix or data frame (u and v, the change along the x and y axes, respectively)}
}
\value{
The correlation coefficient and a probability value
}
\description{
Calculates a correlation coefficient and p-value for two independent sets of vectors.
}
\examples{
x <- matrix(rnorm(200, 2, 1), ncol=2)
y <- matrix(rnorm(200, 3, 1), ncol=2)

vector_corr(x, y)
}

